requirements.txt
openai>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.65.0
python-dotenv>=0.19.0
tenacity>=8.0.1
rich>=10.0.0
pydantic>=2.0.0
datasets>=2.15.0
huggingface-hub>=0.19.0
typer>=0.9.0
PyYAML>=6.0.1
